feat: refactor RAG service by modularizing components and improving rate limiting
